# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Sales Predictor API
  version: 1.0.0
  description: API for sales forecasting, inventory management, and market basket analysis
servers:
  - url: http://localhost:5000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          description: User ID
          example: "507f1f77bcf86cd799439011"
        username:
          type: string
          description: Unique username
          example: "john_doe"
        email:
          type: string
          description: Unique email
          example: "john.doe@example.com"
        role:
          type: string
          enum: [Manager, Planner, Owner, Admin]
          description: User role
          example: "Manager"
      required: [username, email, role]
    Sale:
      type: object
      properties:
        id:
          type: string
          description: Sale ID
          example: "507f191e810c19729de860ea"
        userId:
          type: string
          description: User who recorded the sale
          example: "507f1f77bcf86cd799439011"
        date:
          type: string
          format: date-time
          description: Sale date
          example: "2025-05-20T14:30:00Z"
        totalAmount:
          type: number
          description: Total sale amount
          example: 150.75
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
                description: Product ID
                example: "507f191e810c19729de860eb"
              quantity:
                type: integer
                description: Quantity sold
                example: 2
            required: [productId, quantity]
          example:
            - productId: "507f191e810c19729de860eb"
              quantity: 2
            - productId: "507f191e810c19729de860ec"
              quantity: 1
        promotion:
          type: boolean
          description: Whether sale involved a promotion
          example: true
      required: [userId, date, totalAmount, items]
...
paths:
...
  /api/sales/upload:
    post:
      summary: Upload sales CSV
      tags: [Sales]
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV file with sales data
              required: [file]
      responses:
        "201":
          description: Sales uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Sales uploaded"
                  sales:
                    type: array
                    items:
                      $ref: "#/components/schemas/Sale"
              example:
                message: "Sales uploaded"
                sales:
                  - id: "507f191e810c19729de860ea"
                    userId: "507f1f77bcf86cd799439011"
                    date: "2025-05-20T14:30:00Z"
                    totalAmount: 150.75
                    items:
                      - productId: "507f191e810c19729de860eb"
                        quantity: 2
                      - productId: "507f191e810c19729de860ec"
                        quantity: 1
                    promotion: true
        "400":
          description: Invalid CSV or product not found
          content:
            application/json:
              example:
                error: "Invalid CSV format"
        "401":
          description: Unauthorized
          content:
            application/json:
              example:
                error: "Unauthorized"
        "403":
          description: Access denied (non-Manager)
          content:
            application/json:
              example:
                error: "Access denied"
  /api/sales:
    get:
      summary: List sales
      tags: [Sales]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
            format: date-time
            example: "2025-05-01T00:00:00Z"
          description: Start date filter
        - in: query
          name: endDate
          schema:
            type: string
            format: date-time
            example: "2025-05-31T23:59:59Z"
          description: End date filter
        - in: query
          name: promotion
          schema:
            type: boolean
            example: true
          description: Promotion filter
      responses:
        "200":
          description: Sales list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
              example:
                - id: "507f191e810c19729de860ea"
                  userId: "507f1f77bcf86cd799439011"
                  date: "2025-05-20T14:30:00Z"
                  totalAmount: 150.75
                  items:
                    - productId: "507f191e810c19729de860eb"
                      quantity: 2
                    - productId: "507f191e810c19729de860ec"
                      quantity: 1
                  promotion: true
                - id: "507f191e810c19729de860ef"
                  userId: "507f1f77bcf86cd799439011"
                  date: "2025-05-21T10:15:00Z"
                  totalAmount: 89.50
                  items:
                    - productId: "507f191e810c19729de860ed"
                      quantity: 1
                  promotion: false
        "401":
          description: Unauthorized
          content:
            application/json:
              example:
                error: "Unauthorized"
  /api/sales/{id}:
    get:
      summary: Get sale details
      tags: [Sales]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: "507f191e810c19729de860ea"
          description: Sale ID
      responses:
        "200":
          description: Sale details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
              example:
                id: "507f191e810c19729de860ea"
                userId: "507f1f77bcf86cd799439011"
                date: "2025-05-20T14:30:00Z"
                totalAmount: 150.75
                items:
                  - productId: "507f191e810c19729de860eb"
                    quantity: 2
                  - productId: "507f191e810c19729de860ec"
                    quantity: 1
                promotion: true
        "404":
          description: Sale not found
          content:
            application/json:
              example:
                error: "Sale not found"
        "401":
          description: Unauthorized
          content:
            application/json:
              example:
                error: "Unauthorized"
    put:
      summary: Update sale
      tags: [Sales]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: "507f191e810c19729de860ea"
          description: Sale ID
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                  format: date-time
                  example: "2025-05-20T15:00:00Z"
                totalAmount:
                  type: number
                  example: 175.25
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      productId:
                        type: string
                        example: "507f191e810c19729de860eb"
                      quantity:
                        type: integer
                        example: 3
                  example:
                    - productId: "507f191e810c19729de860eb"
                      quantity: 3
                    - productId: "507f191e810c19729de860ec"
                      quantity: 1
                promotion:
                  type: boolean
                  example: false
      responses:
        "200":
          description: Sale updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
              example:
                id: "507f191e810c19729de860ea"
                userId: "507f1f77bcf86cd799439011"
                date: "2025-05-20T15:00:00Z"
                totalAmount: 175.25
                items:
                  - productId: "507f191e810c19729de860eb"
                    quantity: 3
                  - productId: "507f191e810c19729de860ec"
                    quantity: 1
                promotion: false
        "404":
          description: Sale not found
          content:
            application/json:
              example:
                error: "Sale not found"
        "400":
          description: Validation error
          content:
            application/json:
              example:
                errors:
                  - msg: "Invalid total amount"
                    param: "totalAmount"
        "401":
          description: Unauthorized
          content:
            application/json:
              example:
                error: "Unauthorized"
        "403":
          description: Access denied (non-Manager)
          content:
            application/json:
              example:
                error: "Access denied"
    delete:
      summary: Delete sale
      tags: [Sales]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: "507f191e810c19729de860ea"
          description: Sale ID
      responses:
        "200":
          description: Sale deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Sale deleted"
              example:
                message: "Sale deleted"
        "404":
          description: Sale not found
          content:
            application/json:
              example:
                error: "Sale not found"
        "401":
          description: Unauthorized
          content:
            application/json:
              example:
                error: "Unauthorized"
        "403":
          description: Access denied (non-Manager)
          content:
            application/json:
              example:
                error: "Access denied"
...
  /api/users:
    post:
      summary: Create a new user
      tags: [Users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "alice_jones"
                email:
                  type: string
                  example: "alice.jones@example.com"
                password:
                  type: string
                  example: "SecurePass789!"
                role:
                  type: string
                  enum: [Manager, Planner, Owner, Admin]
                  example: "Planner"
              required: [username, email, password, role]
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User created"
                  user:
                    $ref: "#/components/schemas/User"
              example:
                message: "User created"
                user:
                  id: "507f1f77bcf86cd799439012"
                  username: "alice_jones"
                  email: "alice.jones@example.com"
                  role: "Planner"
        "400":
          description: Validation error or username/email taken
          content:
            application/json:
              example:
                error: "Username or email already taken"
        "401":
          description: Unauthorized
          content:
            application/json:
              example:
                error: "Unauthorized"
        "403":
          description: Access denied (non-Admin)
          content:
            application/json:
              example:
                error: "Access denied"
    get:
      summary: List users
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: role
          schema:
            type: string
            enum: [Manager, Planner, Owner, Admin]
            example: "Manager"
          description: Filter by role
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
              example:
                - id: "507f1f77bcf86cd799439011"
                  username: "john_doe"
                  email: "john.doe@example.com"
                  role: "Manager"
                - id: "507f1f77bcf86cd799439012"
                  username: "alice_jones"
                  email: "alice.jones@example.com"
                  role: "Manager"
        "401":
          description: Unauthorized
          content:
            application/json:
              example:
                error: "Unauthorized"
        "403":
          description: Access denied (non-Admin)
          content:
            application/json:
              example:
                error: "Access denied"
  /api/users/{id}:
    put:
      summary: Update user
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: "507f1f77bcf86cd799439011"
          description: User ID
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "john_doe_updated"
                email:
                  type: string
                  example: "john.updated@example.com"
                password:
                  type: string
                  example: "NewPass789!"
                role:
                  type: string
                  enum: [Manager, Planner, Owner, Admin]
                  example: "Admin"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User updated"
                  user:
                    $ref: "#/components/schemas/User"
              example:
                message: "User updated"
                user:
                  id: "507f1f77bcf86cd799439011"
                  username: "john_doe_updated"
                  email: "john.updated@example.com"
                  role: "Admin"
        "404":
          description: User not found
          content:
            application/json:
              example:
                error: "User not found"
        "400":
          description: Validation error
          content:
            application/json:
              example:
                errors:
                  - msg: "Valid email is required"
                    param: "email"
        "401":
          description: Unauthorized
          content:
            application/json:
              example:
                error: "Unauthorized"
        "403":
          description: Access denied (non-Admin)
          content:
            application/json:
              example:
                error: "Access denied"
    delete:
      summary: Delete user
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: "507f1f77bcf86cd799439011"
          description: User ID
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User deleted"
              example:
                message: "User deleted"
        "404":
          description: User not found
          content:
            application/json:
              example:
                error: "User not found"
        "401":
          description: Unauthorized
          content:
            application/json:
              example:
                error: "Unauthorized"
        "403":
          description: Access denied (non-Admin)
          content:
            application/json:
              example:
                error: "Access denied"